requirements.txt
numpy
pyyaml
python-dotenv
pillow
IPython
shapely
moviepy
networkx
cairosvg
matplotlib
transformers
diffusers
accelerate
jaxtyping
imageio[pyav]
scikit-image
omegaconf
opencv-python
BeautifulSoup4
visdom
triton
numba
scipy
timm
scikit-fmm
einops
safetensors
datasets
scikit-learn
kornia